+++ audio.js
@@ -6,16 +6,41 @@ window.addEventListener("DOMContentLoaded", () => {
 	let closingDown = false;
 
 	// Initialize the recorder
-	var recorder = new WWSUrecorder("assets/plugins/mediarecorder/mp3-mediarecorder-worker.js", window.settings.deviceId);
-	recorder.on("recorderReady", "recorder", () => {
-		window.ipc.renderer.send("console", ["log", "Recorder: Process is ready"]);
-		window.ipc.renderer.send("recorderReady", []);
-	});
+	var recorder = new WWSUrecorder(
+		"assets/plugins/mediarecorder/mp3-mediarecorder-worker.js"
+	);
+
+	// TODO: Initialize the silence detection
+	var silence = new WWSUsilence(window.settings);
+
+	// TODO: Initialize the skywayjs remote broadcasting
+	var remote = new WWSUsilence(window.settings);
+
+	// Initialize the audio manager
+	var audioManager = new WWSUAudioManager(
+		window.settings,
+		window.saveSettings,
+		recorder,
+		silence,
+		remote
+	);
+
+	window.ipc.renderer.send("console", ["log", "Audio: Process is ready"]);
+	window.ipc.renderer.send("audioReady", []);
+
+
+
+	/**
+	 *	RECORDER
+	 */
+
+
+
 	recorder.on("recorderStopped", "recorder", (file) => {
-		console.log(`Recorder: Recording ${file} ended.`);
+		console.log(`Audio: Recording ${file} ended.`);
 		window.ipc.renderer.send("console", [
 			"log",
-			`Recorder: Recording ${file} ended.`,
+			`Audio: Recording ${file} ended.`,
 		]);
 
 		// Close the process if we are pending closing and no file was returned (aka no file to save).
@@ -24,35 +49,37 @@ window.addEventListener("DOMContentLoaded", () => {
 		}
 	});
 	recorder.on("recorderStarted", "recorder", (file) => {
-		console.log(`Recorder: Recording ${file} started.`);
+		console.log(`Audio: Recording ${file} started.`);
 		window.ipc.renderer.send("console", [
 			"log",
-			`Recorder: Recording ${file} started.`,
+			`Audio: Recording ${file} started.`,
 		]);
 	});
 
 	// Pass volume info to main process to be passed to renderer
+	// TODO
 	recorder.audio.on("audioVolume", "recorder", (volume) => {
 		window.ipc.renderer.send("recorderVolume", [volume]);
 	});
 
 	// Pass encoded info to main process to be saved
 	recorder.on("recorderEncoded", "recorder", (file, reader) => {
-		console.log(`Recorder: Recording ${file} finished encoding.`);
+		console.log(`Audio: Recording ${file} finished encoding.`);
 		window.ipc.renderer.send("console", [
 			"log",
-			`Recorder: Recording ${file} finished encoding.`,
+			`Audio: Recording ${file} finished encoding.`,
 		]);
 		window.ipc.main.send("recorderEncoded", [file, reader]);
 	});
 
 	// Listen for device change requests
+	// TODO
 	window.ipc.on("recorderChangeDevice", (event, arg) => {
-		console.log(`Recorder: Changing recording device to ${arg[0]}`);
+		console.log(`Audio: Changing recording device to ${arg[0]}`);
 		recorder.audio.changeDevice(arg[0]);
 		window.ipc.renderer.send("console", [
 			"log",
-			`Recorder: Changed recording device to ${arg[0]}`,
+			`Audio: Changed recording device to ${arg[0]}`,
 		]);
 	});
 
@@ -61,7 +88,7 @@ window.addEventListener("DOMContentLoaded", () => {
 		console.log(`Audio file saved: ${arg[0]}`);
 		window.ipc.renderer.send("console", [
 			"log",
-			`Recorder: Audio file saved to ${arg[0]}`,
+			`Audio: Audio file saved to ${arg[0]}`,
 		]);
 
 		// Close the process if we are pending closing
@@ -75,7 +102,7 @@ window.addEventListener("DOMContentLoaded", () => {
 		recorder.newRecording(arg[0], arg[1] || window.settings.delay);
 		window.ipc.renderer.send("console", [
 			"log",
-			`Recorder: Recording ${arg[0]} will start in ${
+			`Audio: Recording ${arg[0]} will start in ${
 				arg[1] || window.settings.delay
 			} milliseconds.`,
 		]);
@@ -86,7 +113,7 @@ window.addEventListener("DOMContentLoaded", () => {
 		recorder.stopRecording(arg[0] || window.settings.delay);
 		window.ipc.renderer.send("console", [
 			"log",
-			`Recorder: Recording ${arg[0]} will stop in ${
+			`Audio: Recording ${arg[0]} will stop in ${
 				arg[0] || window.settings.delay
 			} milliseconds.`,
 		]);
@@ -94,7 +121,9 @@ window.addEventListener("DOMContentLoaded", () => {
 
 	window.ipc.on("shutDown", (arg) => {
 		closingDown = true;
-		console.log(`Recorder: shut down requested.`)
+		console.log(`Audio: shut down requested.`);
 		recorder.stopRecording(-1);
 	});
+
+	
 });