+++ changelog.md
@@ -2,15 +2,23 @@
 
 ## 8.0.0-alpha.18
 
+### Removed
+
+- recorder.deviceId and silence.deviceId; to be replaced with audio array, each containing recorder and silence booleans whether or not they should be responsible for silence detection or recording.
+
 ### Added
+
 - Support for timesheet notes
+- Schema validation to config
 
 ## 8.0.0-alpha.17 - 2020-10-16
 
 ### Removed
+
 - Buttons in notifications; they do not work with Electron IPC.
 
 ### Added
+
 - Notification when the main UI of DJ Controls crashes.
 - Informative messages on certain pages.
 - Helper messages for check in due date and check in/out quantity fields in the inventory system.
@@ -19,36 +27,44 @@
 - Support for master director (ID = 0): cannot be removed, disabled as admin, nor edited by anyone other than the master director.
 
 ### Changed
+
 - Wording in the event conflicts window to be easier to read.
 
 ## 8.0.0-alpha.16 - 2020-10-13
 
 ### Added
+
 - Delay system accountability logging
 - Hosts management
 
 ### Changed
+
 - Using CSS box shadow flashing for nav icons instead of flashing color class, which was unreliable.
 
 ### Updated
+
 - Packages and plugins
 
 ## 8.0.0-alpha.15
 
 ### Added
+
 - Inventory management and check in/out system.
 
 ## 8.0.0-alpha.14 - 2020-09-24
 
 ### Added
+
 - Director management
 
 ### Changed
+
 - Confirmation dialogs now use iziModal and Alpaca instead of iziToast as large iziToast messages do not scroll on mobile / small screens.
 - Authorization prompts also use iziModal and Alpaca instead of iziToast.
 - Meta info background color on the dashboard reflects the type of broadcast / state of WWSU.
 
 ### Fixed
+
 - Changelog file had improper header levels for changed/added/removed/etc, which should have been h3 but were instead h2.
 - Admin Director authorization was showing all directors in the dropdown selection instead of only admin directors; this was fixed.
 - Open iziModals were on top of "connection lost" etc overlays when the overlays should have been on top of / blocking the modals.
@@ -56,22 +72,27 @@
 ## 8.0.0-alpha.13 - 2020-09-20
 
 ### Updated
+
 - AdminLTE and plugins to latest versions
 - NPM packages
 
 ### Changed
+
 - Now using WCAG AAA compliant colors both in AdminLTE and in calendar colors
 - Calendar event text will turn black if event background color is light so it can be read
 
 ## 8.0.0-alpha.12 - 2020-09-17
 
 ### Removed
+
 - TinyMCE autosave plugin no longer used; caused a bug preventing DJ Controls from closing.
 
 ### Changed
+
 - Installer packages. Now, MacOS will use .pkg (instead of .dmg) and Linux will use .deb or .rpm. Windows will still use .exe, but the installer now allows to install on entire machine.
 
 ### Fixed
+
 - TinyMCE blocking DJ Controls from closing when clicking the close button (caused by TinyMCE autosave plugin; which we disabled).
 - Timesheet edit bug which always required Clocked time in when it was supposed to only require that for certain approval statuses.
 - Reduced padding on administration To-do menu item notification numbers so they do not run into the To-do text.
@@ -80,30 +101,37 @@
 ## 8.0.0-alpha.11 - 2020-09-15
 
 ### Added
+
 - Timesheets tab for viewing and modifying director timesheets
 
 ### Changed
+
 - Some of the language in the update available dialog.
 
 ## 8.0.0-alpha.10 - 2020-09-08
 
 ### Added
+
 - Notifications when a new version of DJ Controls is available (uses WWSU API to check for updates instead of autoUpdater, which does not work).
 
 ## 8.0.0-alpha.9 - 2020-09-06
 
 ### Removed
+
 - WebAudioRecorder.js as it no longer works; replacing with browser MediaRecorder.
 
 ### Added
+
 - MP3 program recorder
 
 ## 8.0.0-alpha.8 - 2020-09-06
 
 ### Removed
+
 - Auto update checking; still does not work and is taking more time than it is worth to try and get working.
 
 ### Updated
+
 - Travis cache
 
 ## 8.0.0-alpha.7 - 2020-09-06
@@ -111,9 +139,11 @@
 THIS IS A FAILED RELEASE. DO NOT USE.
 
 ### Added
+
 - Activated update notification functionality (UNTESTED).
 
 ### Updated
+
 - electron
 - electron-store
 - electron-builder